requirements.txt
gTTS==2.3.2
pydub==0.25.1
moviepy==1.0.3
pillow==10.0.0
openai==1.5.0
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.93.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
streamlit==1.27.0 